requirements.txt
requests==2.31.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-telegram-bot
pytz

